# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Progetto Gruppo T31 - IS2022",
        "description": "API del progetto del Gruppo T31 svolto durente il corso di Ingegneria del software - Anno Accademico 2022/23",
        "contact": {
            "name": "Luca Dematté - Referente gruppo T31",
            "email": "luca.dematte-2@studenti.unitn.it"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "gruppo-t31-api.onrender.com",
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "paths": {
...
        "/prenotazione": {
            "post": {
                "tags": ["prenotazione"],
                "summary": "Creazione di una nuova prenotazione",
                "description": "NON utilizzabile per test a causa di un bug di SwaggerUI",
                "operationId": "nuovaPrenotazione",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "proprietario",
                        "in": "formData",
                        "description": "ID dell'utente che effettua la prenotazione",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "eventoCollegato",
                        "in": "formData",
                        "description": "ID dell'evento collegato alla prenotazione",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "ricorrenze",
                        "in": "formData",
                        "description": "Lista di ricorrenze legate a questa prenotazione",
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "inizio": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "Data e ora di inizio della ricorrenza"
                                },
                                "fine": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "Data e ora di fine della ricorrenza"
                                },
                                "spaziPrenotati": {
                                    "type": "array",
                                    "items": { "type": "string" },
                                    "description": "Lista di ID degli spazi utilizzati in questa ricorrenza"
                                },
                                "serviziPrenotati": {
                                    "type": "array",
                                    "items": { "type": "string" },
                                    "description": "Lista di ID dei servizi utilizzati in questa ricorrenza"
                                }
                            }
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created - Prenotazione effettuata correttamente",
                        "schema": { "$ref": "#/definitions/Prenotazione" }
                    },
                    "400": {
                        "description": "Bad request - I parametri richiesti non sono specificati",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
...
        "/prenotazione/byID": {
            "get": {
                "tags": ["prenotazione"],
                "summary": "Ottenere una prenotazione tramite il suo ID",
                "description": "",
                "operationId": "getPrenotazioneConID",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID della prenotazione da restituire",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/Prenotazione" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Una prenotazione con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/prenotazione/ricorrenze": {
            "get": {
                "tags": ["prenotazione"],
                "summary": "Ottenere la lista delle ricorrenze legate a una prenotazione",
                "description": "",
                "operationId": "getRicorrenzePrenotazione",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID della prenotazione",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/Ricorrenza" }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Una prenotazione con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/prenotazione/{id}": {
            "patch": {
                "tags": ["prenotazione"],
                "summary": "Modifica di una prenotazione",
                "description": "",
                "operationId": "modificaPrenotazione",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID della prenotazione da modificare",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "proprietario",
                        "in": "formData",
                        "description": "ID dell'utente che effettua la prenotazione",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "statusPagamento",
                        "in": "formData",
                        "description": "Stato del pagamento: 0 -> da pagare, 1 -> pagato online, 2 -> pagato in contanti",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "name": "importoPagamento",
                        "in": "formData",
                        "description": "Totale da pagare",
                        "type": "number",
                        "format": "float",
                        "required": false
                    },
                    {
                        "name": "eventoCollegato",
                        "in": "formData",
                        "description": "ID dell'evento collegato alla prenotazione",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "ricorrenze",
                        "in": "formData",
                        "description": "Lista di ricorrenze legate a questa prenotazione",
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "inizio": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "Data e ora di inizio della ricorrenza"
                                },
                                "fine": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "Data e ora di fine della ricorrenza"
                                },
                                "spaziPrenotati": {
                                    "type": "array",
                                    "items": { "type": "string" },
                                    "description": "Lista di ID degli spazi utilizzati in questa ricorrenza",
                                    "example": ""
                                },
                                "serviziPrenotati": {
                                    "type": "array",
                                    "items": { "type": "string" },
                                    "description": "Lista di ID dei servizi utilizzati in questa ricorrenza",
                                    "example": ""
                                }
                            }
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Prenotazione modificata correttamente",
                        "schema": { "$ref": "#/definitions/Prenotazione" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Una prenotazione con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            },
            "delete": {
                "tags": ["prenotazione"],
                "summary": "Eliminazione di una prenotazione e delle sue ricorrenze",
                "description": "",
                "operationId": "eliminaPrenotazione",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID della prenotazione da eliminare",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Eliminazione della prenotazione effettuata correttamente",
                        "schema": { "$ref": "#/definitions/Prenotazione" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Una prenotazione con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Spazio": {
            "type": "object",
            "required": ["_id", "nome", "tipologia", "prezzoIniziale", "prezzoOra"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco dello spazio"
                },
                "nome": {
                    "type": "string",
                    "description": "Nome dello spazio",
                    "example": "Sala polifunzionale"
                },
                "descrizione": {
                    "type": "string",
                    "description": "Breve descrizione dello spazio",
                    "example": "Aula molto grande (100 posti) per spettacoli, conferenze, ..."
                },
                "tipologia": {
                    "type": "string",
                    "description": "Tipo di spazio",
                    "enum": ["Locale interno", "Locale attrezzato", "Spazio esterno"],
                    "example": "Locale interno"
                },
                "prezzoIniziale": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo dello spazio quando viene utilizzato",
                    "example": 10.5
                },
                "prezzoOra": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo dello spazio per ogni ora di utilizzo",
                    "example": 4.25
                },
                "URLfoto": {
                    "type": "string",
                    "description": "Percorso della foto dello spazio",
                    "example": "/images/spazi/default.png"
                },
                "servizi": {
                    "type": "array",
                    "description": "Lista dei servizi che possono essere utilizzati in questo spazio",
                    "items": { "$ref": "#/definitions/Servizio" },
                    "required": false
                }
            }
        },
        "Servizio": {
            "type": "object",
            "required": ["_id", "nome", "tipologia", "prezzoIniziale", "prezzoOra"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco del servizio"
                },
                "nome": {
                    "type": "string",
                    "description": "Nome del servizio",
                    "example": "Proiettore"
                },
                "descrizione": {
                    "type": "string",
                    "description": "Breve descrizione del servizio",
                    "example": "Proiettore 720p con interfaccia HDMI e VGA"
                },
                "tipologia": {
                    "type": "string",
                    "description": "Tipo di servizio",
                    "enum": ["Arredamento", "Attrezzatura specifica", "Altro"],
                    "example": "Attrezzatura specifica"
                },
                "prezzoIniziale": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo del servizio quando viene utilizzato",
                    "example": 2.5
                },
                "prezzoOra": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo del servizio per ogni ora di utilizzo",
                    "example": 1.0
                },
                "URLfoto": {
                    "type": "string",
                    "description": "Percorso della foto del servizio",
                    "example": "/images/servizi/default.png"
                }
            }
        },
        "Prenotazione": {
            "type": "object",
            "required": [
                "_id",
                "proprietario",
                "statusPagamento",
                "importoPagamento",
                "ricorrenze"
            ],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco della prenotazione"
                },
                "proprietario": {
                    "type": "string",
                    "description": "ID dell'utente che effettua la prenotazione"
                },
                "statusPagamento": {
                    "type": "integer",
                    "description": "Stato del pagamento: 0 -> da pagare, 1 -> pagato online, 2 -> pagato in contanti",
                    "example": 1
                },
                "importoPagamento": {
                    "type": "number",
                    "format": "float",
                    "description": "Totale da pagare",
                    "example": 30.75
                },
                "eventoCollegato": {
                    "type": "string",
                    "description": "ID dell'evento collegato alla prenotazione"
                },
                "ricorrenze": {
                    "type": "array",
                    "items": { "$ref": "#/definitions/Ricorrenza" },
                    "description": "Lista di ricorrenze legate a questa prenotazione"
                }
            }
        },
        "Ricorrenza": {
            "type": "object",
            "required": ["_id", "inizio", "fine", "spaziPrenotati"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco della ricorrenza"
                },
                "inizio": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Data e ora di inizio della ricorrenza",
                    "example": "2023-01-10T16:30:00.000Z"
                },
                "fine": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Data e ora di fine della ricorrenza",
                    "example": "2023-01-10T18:30:00.000Z"
                },
                "spaziPrenotati": {
                    "type": "array",
                    "items": { "$ref": "#/definitions/Spazio" },
                    "description": "Lista degli spazi utilizzati in questa ricorrenza"
                },
                "serviziPrenotati": {
                    "type": "array",
                    "items": { "$ref": "#/definitions/Servizio" },
                    "description": "Lista dei servizi utilizzati in questa ricorrenza"
                }
            }
        },
        "ErroreAPI": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
                "code": {
                    "type": "integer",
                    "description": "Codice di risposta HTTP"
                },
                "message": {
                    "type": "string",
                    "description": "Messaggio di testo sull'errore"
                }
            }
        }
    }
}